Support for additional parameters
Cf http://php.net/manual/en/function.mail.php

The additional_parameters parameter can be used to pass additional flags as command line options to the program configured to be used when sending mail, as defined by the sendmail_path configuration setting. For example, this can be used to set the envelope sender address when using sendmail with the -f sendmail option.

Signed-off-by: Matthieu Baudoux <matthieu@hypernovae.be>